Reduces refresh rate pending messages

+++ src/test/groovy/com/leanforge/game/TestApplication.groovy
@@ -59,7 +59,7 @@ class TestApplication {
         }
 
         @Override
-        int count() {
+        long count() {
             return 0
         }
